// packages/cosmos-amino/src/types/types.ts
export interface AminoConverter {
  readonly aminoType: string;
  readonly toAmino: (value: any) => any;
  readonly fromAmino: (value: any) => any;
}

export type TypeUrl = string;
export type AminoConverters = Record<TypeUrl, AminoConverter>;
export type Registry = Array<[TypeUrl, TelescopeGeneratedType]>;

export interface Generated extends TelescopeGeneratedType {
  amino?: AminoConverter;
}